+++ scripts/macros/apply-effect.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyGrappledEffect(tok.actor, this.uuid, 13);
-}