+++ scripts/macros/effects/apply-charmed.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyCharmedEffect(tok.actor, this.uuid);
-}