+++ scripts/macros/effects/apply-exhaustion.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyExhaustionEffect(tok.actor, this.uuid, 1);
-}