+++ scripts/macros/effects/apply-frightened.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyFrightenedEffect(tok.actor, this.uuid);
-}