+++ scripts/macros/effects/apply-petrified.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyPetrifiedEffect(tok.actor, this.uuid);
-}