+++ scripts/macros/effects/apply-restrained.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyRestrainedEffect(tok.actor, this.uuid, 13);
-}