+++ scripts/macros/effects/apply-stunned.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyStunnedEffect(tok.actor, this.uuid);
-}