+++ scripts/macros/effects/apply-unconscious.js
@@ -1,3 +1,0 @@
-for (let tok of canvas.tokens.controlled) {
-    await HomebrewEffects.applyUnconsciousEffect(tok.actor, this.uuid);
-}